python/requirements.txt
pytest
pytest-cov
pyyaml
testfixtures
pdbpp
zope.component
